+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 13.2
 * Add `org.sonar.api.issue.IssueStatus.IN_SANDBOX` enum value
 * Add `org.sonar.api.issue.Issue.STATUS_IN_SANDBOX` constant (deprecated immediately)
+* Introduce new security standards for OWASP LLM 2025
 
 ## 13.1
 * Introduce support for internal tags on issues: